+++ netpyne/analysis/__init__.py
@@ -65,4 +65,4 @@ from .rxd import plotRxDConcentration
 # Import HNN-related functions
 # -------------------------------------------------------------------------------------------------------------------
 from .interactive import iplotDipole, iplotDipoleSpectrogram, iplotDipolePSD, iplotRaster, iplotSpikeHist, \
-    iplotRatePSD, iplotTraces, iplotLFP, iplotConn, iplotRxDConcentration
+    iplotRatePSD, iplotTraces, iplotLFP, iplotConn, iplotRxDConcentration, iplot2Dnet